requirements.txt
alembic==1.17.1
asyncpg==0.30.0
authx==1.4.3
bcrypt==3.2.2
fastapi==0.120.1
fastapi-async-sqlalchemy==0.6.1
greenlet==3.2.1
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.12.3
pydantic[email]==2.12.3
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
python-multipart
pytz==2025.2
pyjwt==2.9.0
requests
sqlalchemy==2.0.44
uvicorn[standard]==0.38.0